requirements.txt

fastapi
uvicorn[standard]
python-multipart
moviepy
ffmpeg-python
openai-whisper
transformers
sentencepiece
torch
soundfile
aiofiles
